refactor : 각 클래스별 주석 제거 및 Impl 패턴 제거
